+++ smart_city_socket_server/traffic_control/tc_command_handler.c
@@ -16,7 +16,7 @@ void tc_id_issuance (void *data)
 void tc_barricade_handler (void *data)
 {
     printf("바리케이드 서브 커맨드: %d\n", *((int *)data));
-    uint8_t data_pkt = *((uint8_t *)data);
+    uint8_t data_pkt = *((int *)data);
 
     if (data_pkt == ON)
     {
@@ -31,7 +31,7 @@ void tc_barricade_handler (void *data)
 void tc_lift_handler (void *data)
 {
     printf("리프트 서브 커맨드: %d\n", *((int *)data));
-    uint8_t data_pkt = *((uint8_t *)data);
+    uint8_t data_pkt = *((int *)data);
 
     if (data_pkt == ON)
     {
@@ -46,7 +46,7 @@ void tc_lift_handler (void *data)
 void tc_street_lamp_handler (void *data)
 {
     printf("가로등 서브 커맨드: %d\n", *((int *)data));
-    uint8_t data_pkt = *((uint8_t *)data);
+    uint8_t data_pkt = *((int *)data);
 
     if (data_pkt == ON)
     {
@@ -61,7 +61,7 @@ void tc_street_lamp_handler (void *data)
 void tc_traffic_light_handler (void *data)
 {
     printf("신호등 서브 커맨드: %d\n", *((int *)data));
-    uint8_t data_pkt = *((uint8_t *)data);
+    uint8_t data_pkt = *((int *)data);
 
     if (data_pkt == ON)
     {
